backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

spacy>=3.7.2
nltk>=3.8.1

opencv-python-headless>=4.9.0
numpy>=1.26.0

python-multipart>=0.0.6

pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.39.0
Pillow>=10.2.0
